python_crawler/requirements.txt
requests==2.28.2
beautifulsoup4==4.11.1
pymysql==1.0.2